+++ 01-cli/src/main/kotlin/ru/spbau/bogomolov/Main.kt
@@ -1,7 +1,7 @@
 package ru.spbau.bogomolov
 
-fun main(args : Array<String>): Unit {
-    val appInstance = CommandLineExecutor()
-    val executor = BashLikeExecutor()
-    appInstance.run(executor)
+fun main(args : Array<String>) {
+    val appInstance = CommandLineInterpretator()
+    val interpretator = BashLikeExecutor()
+    appInstance.run(interpretator)
 }
